started object detection adn pose
