+++ action_planning/package.xml
@@ -10,6 +10,7 @@
   <buildtool_depend>ament_cmake</buildtool_depend>
 
   <depend>rclcpp</depend>
+  <depend>rclpy</depend>
   <depend>moveit_action</depend>
   <depend>geometry_msgs</depend>
   <depend>visualization_msgs</depend>
